ipopt package: set compiler flags in configure (#6714)
This makes use of the new flag_handler logic from d49c7051 to set
compiler flags for ipopt by passing them as arguments to the build
system rather than injecting them into the compiler wrappers. This
avoids conflicts between flags that are chosen by the build system
and flags that are set by the user.